Initial LSP setup
Errors are showing up in C#, JSON, YAML and Lua files. Code suggestions
also seem to work in Lua, will have to figure out how to do this for
Omnisharp.

Also replaced mini.pairs with nvim-autopairs because the latter is smart
enough to not double quotes when your cursor is right before a word!