Replace omnisharp with roslyn.nvim
Tested on smaller projects and it seems to be working fine,
but there are a couple of things missing, such as support
for displaying errors inline, at the moment it just shows an E
in the gutter (but this might be fixed with a plugin?).

Also tested this on macOS and I need to do some keybinding fixes
for it, because I might have focused only on Windows, but it's not
that important at the moment.

Also, some spellchecker thingy might be useful.